requirements.txt
python-telegram-bot
telegram_bot_pagination
requests